allow also to run from different locations

+++ starsky/build/helpers/DotnetTestHelper.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Text;
 using build;
 using Microsoft.Extensions.FileSystemGlobbing;
 using Nuke.Common.IO;
@@ -26,7 +25,7 @@ namespace helpers
 			matcher.AddIncludePatterns(new[] { globSearch });
 			PatternMatchingResult result = matcher.Execute(
 				new DirectoryInfoWrapper(
-					new DirectoryInfo(".")));
+					new DirectoryInfo(WorkingDirectory.GetSolutionParentFolder())));
 			return result.Files.Select(p => p.Path).ToList();
 		}
 
@@ -44,22 +43,33 @@ namespace helpers
 				Information($">> TestNetCore is disable due the --no-unit-test flag");
 				return;
 			}
-
-			var projects = GetFiles("./*test/*.csproj");
+			
+			var projects = GetFiles("*test/*.csproj");
+			if ( projects.Count == 0 )
+			{
+				throw new FileNotFoundException("missing tests in *test/*.csproj" );
+			}
+			
 			foreach(var project in projects)
 			{
+				var projectFullPath = Path.Combine(WorkingDirectory.GetSolutionParentFolder(),
+					project);
 				Information("Testing project " + project);
 
-				var testParentPath = Directory.GetParent(project)?.FullName;
+				var testParentPath = Directory.GetParent(projectFullPath)?.FullName;
+				Information("testParentPath " + testParentPath);
 
 				/* clean test results */
-				var testResultsFolder = System.IO.Path.Combine(testParentPath, "TestResults");
+				var testResultsFolder = Path.Combine(testParentPath!, "TestResults");
 				if (DirectoryExists(testResultsFolder))
 				{
 					Information(">> Removing folder => " + testResultsFolder);
 					Directory.Delete(testResultsFolder,true);
 				}
-
+				
+				var runSettingsFile = Path.Combine(WorkingDirectory.GetSolutionParentFolder(), "build.vstest.runsettings");
+				Console.WriteLine("runSettingsFile " + runSettingsFile);
+				
 				// search for: dotnet test
 				DotNetTest(_ => _
 					.SetConfiguration(configuration)
@@ -69,19 +79,25 @@ namespace helpers
 					.SetVerbosity(DotNetVerbosity.Normal)
 					.SetLoggers("trx;LogFileName=test_results.trx")
 					.SetDataCollector("XPlat Code Coverage")
-					.SetSettingsFile("build.vstest.runsettings")
-					.SetProjectFile(project));
+					.SetSettingsFile(runSettingsFile)
+					.SetProjectFile(projectFullPath));
 
 				Information("on Error: search for: Error Message");
-				var coverageEnum = GetFiles("./**/coverage.opencover.xml");
+				var coverageEnum = GetFiles("**/coverage.opencover.xml");
+
+				foreach ( var coverageItem in coverageEnum )
+				{
+					Information("coverageItem: " + coverageItem);
+				}
 
 				// Get the FirstOrDefault() but there is no LINQ here
-				var coverageFilePath =  System.IO.Path.Combine(testParentPath, "netcore-coverage.opencover.xml");
+				var coverageFilePath =  Path.Combine(testParentPath, "netcore-coverage.opencover.xml");
+				Information("next copy: coverageFilePath " + coverageFilePath);
+
 				foreach(var item in coverageEnum)
 				{
-					CopyFile(item, coverageFilePath, FileExistsPolicy.Overwrite);
+					CopyFile(Path.Combine(WorkingDirectory.GetSolutionParentFolder(), item), coverageFilePath, FileExistsPolicy.Overwrite);
 				}
-				Information("CoverageFile " + coverageFilePath);
 
 				if (!FileExists(coverageFilePath)) {
 					throw new FileNotFoundException("CoverageFile missing " + coverageFilePath);
